Change `weight` column type to `DOUBLE PRECISION`, add item search endpoint in `ItemController`, and extend `ItemRepository` with case-insensitive search by name.

+++ src/main/resources/db/migration/V0__Create_Schema.sql
@@ -30,7 +30,7 @@ CREATE TABLE items (
                        rarity VARCHAR(255),
                        stack_size INTEGER,
                        value INTEGER,
-                       weight NUMERIC(38,2),
+                       weight DOUBLE PRECISION,
                        loot_type_id BIGINT
 );
